Separated calls to get info from alerting confix index in MonitorRunner to separate classes
