+++ alerting/src/main/kotlin/com/amazon/opendistroforelasticsearch/alerting/MonitorRunner.kt
@@ -20,7 +20,6 @@ import com.amazon.opendistroforelasticsearch.alerting.alerts.AlertIndices
 import com.amazon.opendistroforelasticsearch.alerting.alerts.moveAlerts
 import com.amazon.opendistroforelasticsearch.alerting.core.JobRunner
 import com.amazon.opendistroforelasticsearch.alerting.core.model.ScheduledJob
-import com.amazon.opendistroforelasticsearch.alerting.core.model.ScheduledJob.Companion.SCHEDULED_JOBS_INDEX
 import com.amazon.opendistroforelasticsearch.alerting.core.model.SearchInput
 import com.amazon.opendistroforelasticsearch.alerting.elasticapi.convertToMap
 import com.amazon.opendistroforelasticsearch.alerting.elasticapi.firstFailureOrNull
@@ -34,6 +33,7 @@ import com.amazon.opendistroforelasticsearch.alerting.model.Alert.State.ACTIVE
 import com.amazon.opendistroforelasticsearch.alerting.model.Alert.State.COMPLETED
 import com.amazon.opendistroforelasticsearch.alerting.model.Alert.State.DELETED
 import com.amazon.opendistroforelasticsearch.alerting.model.Alert.State.ERROR
+import com.amazon.opendistroforelasticsearch.alerting.model.AlertingConfigAccessor
 import com.amazon.opendistroforelasticsearch.alerting.model.InputRunResults
 import com.amazon.opendistroforelasticsearch.alerting.model.Monitor
 import com.amazon.opendistroforelasticsearch.alerting.model.MonitorRunResult
@@ -43,12 +43,7 @@ import com.amazon.opendistroforelasticsearch.alerting.model.action.Action
 import com.amazon.opendistroforelasticsearch.alerting.model.action.Action.Companion.MESSAGE
 import com.amazon.opendistroforelasticsearch.alerting.model.action.Action.Companion.MESSAGE_ID
 import com.amazon.opendistroforelasticsearch.alerting.model.action.Action.Companion.SUBJECT
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.Destination
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.DestinationContext
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.email.Email
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.email.EmailAccount
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.email.EmailGroup
-import com.amazon.opendistroforelasticsearch.alerting.model.destination.email.Recipient
+import com.amazon.opendistroforelasticsearch.alerting.model.destination.DestinationContextFactory
 import com.amazon.opendistroforelasticsearch.alerting.script.TriggerExecutionContext
 import com.amazon.opendistroforelasticsearch.alerting.script.TriggerScript
 import com.amazon.opendistroforelasticsearch.alerting.settings.AlertingSettings.Companion.ALERT_BACKOFF_COUNT
@@ -56,7 +51,6 @@ import com.amazon.opendistroforelasticsearch.alerting.settings.AlertingSettings.
 import com.amazon.opendistroforelasticsearch.alerting.settings.AlertingSettings.Companion.MOVE_ALERTS_BACKOFF_COUNT
 import com.amazon.opendistroforelasticsearch.alerting.settings.AlertingSettings.Companion.MOVE_ALERTS_BACKOFF_MILLIS
 import com.amazon.opendistroforelasticsearch.alerting.settings.DestinationSettings.Companion.loadDestinationSettings
-import com.amazon.opendistroforelasticsearch.alerting.util.DestinationType
 import com.amazon.opendistroforelasticsearch.alerting.util.IndexUtils
 import org.apache.logging.log4j.LogManager
 import kotlinx.coroutines.CoroutineScope
@@ -71,8 +65,6 @@ import org.elasticsearch.action.bulk.BackoffPolicy
 import org.elasticsearch.action.bulk.BulkRequest
 import org.elasticsearch.action.bulk.BulkResponse
 import org.elasticsearch.action.delete.DeleteRequest
-import org.elasticsearch.action.get.GetRequest
-import org.elasticsearch.action.get.GetResponse
 import org.elasticsearch.action.index.IndexRequest
 import org.elasticsearch.action.search.SearchRequest
 import org.elasticsearch.action.search.SearchResponse
@@ -81,7 +73,6 @@ import org.elasticsearch.cluster.service.ClusterService
 import org.elasticsearch.common.Strings
 import org.elasticsearch.common.bytes.BytesReference
 import org.elasticsearch.common.component.AbstractLifecycleComponent
-import org.elasticsearch.common.settings.SecureString
 import org.elasticsearch.common.settings.Settings
 import org.elasticsearch.common.xcontent.LoggingDeprecationHandler
 import org.elasticsearch.common.xcontent.NamedXContentRegistry
@@ -124,6 +115,7 @@ class MonitorRunner(
         BackoffPolicy.exponentialBackoff(MOVE_ALERTS_BACKOFF_MILLIS.get(settings), MOVE_ALERTS_BACKOFF_COUNT.get(settings))
 
     @Volatile private var destinationSettings = loadDestinationSettings(settings)
+    @Volatile private var destinationContextFactory = DestinationContextFactory(client, xContentRegistry, destinationSettings)
 
     init {
         clusterService.clusterSettings.addSettingsUpdateConsumer(ALERT_BACKOFF_MILLIS, ALERT_BACKOFF_COUNT) {
@@ -137,6 +129,9 @@ class MonitorRunner(
     /** Update destination settings when the reload API is called so that new keystore values are visible */
     fun reloadDestinationSettings(settings: Settings) {
         destinationSettings = loadDestinationSettings(settings)
+
+        // Update destinationContextFactory as well since destinationSettings has been updated
+        destinationContextFactory.updateDestinationSettings(destinationSettings)
     }
 
     override fun doStart() {
@@ -447,8 +442,8 @@ class MonitorRunner(
             }
             if (!dryrun) {
                 withContext(Dispatchers.IO) {
-                    val destination = getDestinationInfo(action.destinationId)
-                    val destinationCtx = getDestinationContext(destination)
+                    val destination = AlertingConfigAccessor.getDestinationInfo(client, xContentRegistry, action.destinationId)
+                    val destinationCtx = destinationContextFactory.getDestinationContext(destination)
                     actionOutput[MESSAGE_ID] = destination.publish(
                         actionOutput[SUBJECT],
                         actionOutput[MESSAGE]!!,
@@ -466,109 +461,6 @@ class MonitorRunner(
         return scriptService.compile(template, TemplateScript.CONTEXT)
                 .newInstance(template.params + mapOf("ctx" to ctx.asTemplateArg()))
                 .execute()
-    }
-
-    private suspend fun getDestinationContext(destination: Destination): DestinationContext {
-        var destinationContext = DestinationContext()
-        // Populate DestinationContext based on Destination type
-        if (destination.type == DestinationType.EMAIL) {
-            val email = destination.email
-            // Email should not be null in this case but checking here to make safe calls to the attributes below
-            requireNotNull(email) { "Email in Destination: $destination was null" }
-
-            // Get the EmailAccount information by doc ID
-            var emailAccount = getEmailAccountInfo(email.emailAccountID)
-
-            // Populate the username and password for the EmailAccount if authentication is enabled
-            emailAccount = addEmailCredentials(emailAccount)
-
-            // Get the email recipients as a unique list of email strings since
-            // recipients can be a combination of EmailGroups and single emails
-            val uniqueListOfRecipients = getUniqueListOfEmailRecipients(email)
-
-            destinationContext = destinationContext.copy(emailAccount = emailAccount, recipients = uniqueListOfRecipients)
-        }
-
-        return destinationContext
-    }
-
-    private fun addEmailCredentials(emailAccount: EmailAccount): EmailAccount {
-        // Retrieve and populate the EmailAccount object with credentials if authentication is enabled
-        if (emailAccount.method != EmailAccount.MethodType.NONE) {
-            val emailUsername: SecureString? = destinationSettings[emailAccount.name]?.emailUsername
-            val emailPassword: SecureString? = destinationSettings[emailAccount.name]?.emailPassword
-
-            return emailAccount.copy(username = emailUsername, password = emailPassword)
-        }
-
-        return emailAccount
-    }
-
-    private suspend fun getUniqueListOfEmailRecipients(email: Email): List<String> {
-        val uniqueRecipients: MutableSet<String> = mutableSetOf()
-        email.recipients.forEach { recipient ->
-            when (recipient.type) {
-                // Recipient attributes are checked for being non-null based on type during initialization
-                // so non-null assertion calls are made here
-                Recipient.RecipientType.EMAIL -> uniqueRecipients.add(recipient.email!!)
-                Recipient.RecipientType.EMAIL_GROUP -> {
-                    val emailGroup = getEmailGroupInfo(recipient.emailGroupID!!)
-                    emailGroup.getEmailsAsListOfString().map { uniqueRecipients.add(it) }
-                }
-            }
-        }
-
-        return uniqueRecipients.toList()
-    }
-
-    private suspend fun getDestinationInfo(destinationId: String): Destination {
-        val getRequest = GetRequest(SCHEDULED_JOBS_INDEX, destinationId).routing(destinationId)
-        val getResponse: GetResponse = client.suspendUntil { client.get(getRequest, it) }
-        if (!getResponse.isExists || getResponse.isSourceEmpty) {
-            throw IllegalStateException("Destination document with id $destinationId not found or source is empty")
-        }
-
-        val jobSource = getResponse.sourceAsBytesRef
-        return withContext(Dispatchers.IO) {
-            val xcp = XContentHelper.createParser(xContentRegistry, LoggingDeprecationHandler.INSTANCE,
-                jobSource, XContentType.JSON)
-            ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.nextToken(), xcp::getTokenLocation)
-            ensureExpectedToken(XContentParser.Token.FIELD_NAME, xcp.nextToken(), xcp::getTokenLocation)
-            ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.nextToken(), xcp::getTokenLocation)
-            val destination = Destination.parse(xcp)
-            ensureExpectedToken(XContentParser.Token.END_OBJECT, xcp.nextToken(), xcp::getTokenLocation)
-            destination
-        }
-    }
-
-    private suspend fun getEmailAccountInfo(emailAccountID: String): EmailAccount {
-        val getRequest = GetRequest(SCHEDULED_JOBS_INDEX, emailAccountID).routing(emailAccountID)
-        val getResponse: GetResponse = client.suspendUntil { client.get(getRequest, it) }
-        if (!getResponse.isExists || getResponse.isSourceEmpty) {
-            throw IllegalStateException("Email account document with id $emailAccountID not found or source is empty")
-        }
-
-        val source = getResponse.sourceAsBytesRef
-        return withContext(Dispatchers.IO) {
-            val xcp = XContentHelper.createParser(xContentRegistry, LoggingDeprecationHandler.INSTANCE, source, XContentType.JSON)
-            val emailAccount = EmailAccount.parseWithType(xcp)
-            emailAccount
-        }
-    }
-
-    private suspend fun getEmailGroupInfo(emailGroupID: String): EmailGroup {
-        val getRequest = GetRequest(SCHEDULED_JOBS_INDEX, emailGroupID).routing(emailGroupID)
-        val getResponse: GetResponse = client.suspendUntil { client.get(getRequest, it) }
-        if (!getResponse.isExists || getResponse.isSourceEmpty) {
-            throw IllegalStateException("Email group document with id $emailGroupID not found or source is empty")
-        }
-
-        val source = getResponse.sourceAsBytesRef
-        return withContext(Dispatchers.IO) {
-            val xcp = XContentHelper.createParser(xContentRegistry, LoggingDeprecationHandler.INSTANCE, source, XContentType.JSON)
-            val emailGroup = EmailGroup.parseWithType(xcp)
-            emailGroup
-        }
     }
 
     private fun List<AlertError>?.update(alertError: AlertError?): List<AlertError> {
